% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_subjects.R
\name{associate_subjects}
\alias{associate_subjects}
\title{Find ocurrences in dataframe2 based on equal and tolerance values in dataframe1}
\usage{
associate_subjects(
  data_input1,
  data_input2,
  equal_colname_vector = NULL,
  tolerance_colname_vector = NULL,
  tolerance_value_vector = NULL
)
}
\arguments{
\item{data_input1:}{the first dataframe object to extract subjects to identify}

\item{data_input2:}{the second dataframe object to check for subject identities}

\item{equal_colname_vector:}{vector of column names for variables that must be identical between the dataframes}

\item{tolerance_colname_vector:}{vector of column names for variables that must be similar between the dataframes}

\item{tolerance_value_vector:}{vector of tolerance values (must be same length of "tolerance_colname_vector")}
}
\value{
A list() object with "$data": the merged dataframes for identified subjects; and "$unidentified_data": the subset of dataframe for unidentified subjects (both for "data_input1" and "data_input2")
}
\description{
This function allows to specify column that must have identical values and columns
that must have similar values (with specifide tolerance) between two databases and
it will screen to identify subjects based on passed parameters. IT will return a
dataframe which is the merge of both databases for identified subjects, and a dataframe
which is represented by unidentifed subjects in first dataframe.
}
