% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_multiple_columns.R
\name{value_multiple_columns}
\alias{value_multiple_columns}
\title{Find ocurrences of a specific value in multiple columns}
\usage{
value_multiple_columns(data_input, value, colname_vector)
}
\arguments{
\item{data_input:}{the dataframe input object}

\item{value:}{the value to check for}

\item{colname_vector:}{vector of column names to check}
}
\value{
A vector() object with 0/1 based on absence/presence of specified value in specifed columns for each sample
}
\description{
This function allows to specify a value that must be identified on multiple columns
and it returns a 0/1 vector based on absence/presence of the value
}
